This PR adds support for dyanmic shape parameters, specifically substituting llama's appearance of `config.vocab_size` to `llama.vocab_size`. 

### Motivation
Models that share the same architecture may only differ in the vocab size, making us have to recompile a model library. With dynamic vocab size, we can use the same llama model library for llama-2, code-llama, wizardMath, wizardCoder, etc.

### Changes needed
1. Quantization
a. GroupQuantization: During `_dequantize()`, insetad of passing in `IntImm`, ~~pass in `tir.Var` when shape is not int~~ pass in weight's shape, which is a tir.Var (cannot initialize a new var since it would create `vocab_size_1` in script)
 b. No change is needed for AWQ since it does not seem to quantize the embedding layer nor the `lm_head`: https://github.com/mlc-ai/mlc-llm/blob/73c47628fb921eaab2a12ab43275e79d8f26cd60/python/mlc_chat/compiler/quantization/awq_quantization.py#L119-L121

2. `convert_weight.py`
  a. Modify `_check_param()` such that it allows dynamic shape
  b. ~~`_calc_total_params()` skips dynamic-shape parameters~~ Replace `_calc_total_params()` with `loader.stats.total_param_num`

3. Metadata
  a. For the `param` field in metadata, if dynamic, record ~~`["vocab_size", 4096]`~~ `[-1, 4096]`, rather than previously `[32000, 4096]`
b. ~~Skips the param reading~~ Read in `[-1, 4096]` in runtime `cpp/metadata/model.cc`

Depend on https://github.com/apache/tvm/pull/16284

### Performance
Tested on llama-2 on RTX 4090, virtually no performance degradation.
Before:
```
Statistics: ----------- prefill -----------
throughput: 325.278 tok/s
total tokens: 7 tok
total time: 0.022 s
------------ decode ------------
throughput: 183.725 tok/s
total tokens: 256 tok
total time: 1.393 s
```

After:
```
Statistics: ----------- prefill -----------
throughput: 325.510 tok/s
total tokens: 7 tok
total time: 0.022 s
------------ decode ------------
throughput: 185.398 tok/s
total tokens: 256 tok
total time: 1.381 s
```