@CharlieFRuan 2023-12-11
Hi @junrushao, would really appreciate your suggestions here. Specifically, 2.b and 3.b are not ideal.

For 2.b, I think we could have a `dynamic_shape_map` that gets populated in `_check_shape()`, populating `vocab_size` to either `32000`, `32003`, etc. when comparing `expect` and `actual`. But I am unsure whether it is possible that the quantized weight could have shape like `vocab_size / 4` in the future, a case that I'm unsure what would happen.

For 3.b, I currently do not find the usage of `Param` in `metadata` during runtime. Skipping is definitely not a good option here, but what would be a better way of representing this in runtime? I am not too familiar with tvm runtime, but it seems `ShapeTuple` can only be initialized with concrete values. Perhaps we could populate after loading the actual weight in?