@junrushao 2023-12-14
# Symbolic Parameters

Hey @CharlieFRuan apologize for the delayed response! I was a bit under the weather lately. It's super cool that you figured out almost everything already, and I'd love to share in a few perspectives to help make it more complete!

C1. First of all, about UX, how about supporting symbolic shape Tensor/Parameters in TVM Relax? Instead of writing:

```python
nn.Embedding(tir.Var("vocab_size", "int64"), config.hidden_size)
```

let's allow a more straightforward syntax:

```python
nn.Embedding("vocab_size", config.hidden_size)
```

This can be done by tweaking `[_tensor_placeholder](https://github.com/apache/tvm/blob/fe89ccc360838847f52eb96181ffe69477000835/python/tvm/relax/frontend/nn/core.py#L684)` adding `str` to a possible element type of `shape`. You will likely have to tweak type annotations of a few `nn.Module`s as well. To make things simpler, we could define:

```python
ShapeVar = Union[int, str, tir.PrimExpr]
```

to semantically enumerate the possible types for shape variables.

C2. Handling of symbolic relationship between dynamic shape tensors. There are actually two issues that could potentially cause problems in the future:

- If we print out TVMScript being exported, it's highly possible that we see two distinct variables for `vocab_size`:

```python
@R.function
def prefill(...):
  vocab_size = T.int64()   # for `embed_tokens`
  vocab_size1 = T.int64()  # for `lm_head`
```

It means the underlying TVM compiler is not aware of symbolic relationships like “your `vocab_size` is the same as mine”.

- two `relax.Function` may share the same `tir.Var`. It is conceptually not a big problem, given each function is a relatively independent component in the `IRModule`, but some passes are just assuming otherwise - I was bitten by this issue a couple of times before. In the example below,

```python
@I.ir_module
class Module:

  @R.function
  def prefill(...):
    vocab_size = T.int64() <== A
    ...

  @R.function
  def decode(...):
    vocab_size = T.int64() <== B
    ...
```

TIR variable A and B share the same physical address in memory.

C3. Representing non-static shapes in MLC’s metadata. Currently, if a parameter’s shape is dynamic, use an `is_dynamic` field to replace `shape` when creating metadata. I don’t have much strong opinion, though I would prefer the following:

```python
"shape": [128, -1]
```

where `-1` means dynamic shape. Note that the two approaches both discard symbolic relationship, but my proposal preserves slightly more data on non-dynamic dimensions.

Let me know what you think!