@CharlieFRuan 2023-12-14
Hi @junrushao, thank you so much for the detailed response, and I hope you feel better soon!

I am in line with C1 and C3, and can follow up with the changes; I really look forward to contributing to relax!

For C2 on the other hand, what would be a possible solution?

Besides, I might need ~the end of next week to follow up on this if that is fine -- will probably focus on SLM rollout and documentation revamp first (amidst traveling). Thank you!