@tqchen 2023-12-14
On C2

- First update the nn.Parameter constructor to allow creating a new `tir.Var`, with name "vocab_size"
- Update the relax nn frontend
    -   https://github.com/apache/tvm/blob/unity/python/tvm/relax/frontend/nn/spec.py#L545
    -   Right now notice how we simply pass in a param.shape
    -   Instead we shold do a param.shape remap, that remaps symbolic tir.Var by its name to a new Var when the name first created, and to the same var when we see it in second time.