@tqchen 2024-01-05
@CharlieFRuan please rebase and maybe @junrushao can then take another look. This would be important for us to build SLM models to support all Llama variants